# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Resume Boost API
  version: 1.0.0
  description: API for Resume Boost application, providing resume analysis and job recommendations.
servers:
  - url: http://localhost:8000
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - username
        - createdAt

    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
      required:
        - user
        - token

    LoginCredentials:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    SignupCredentials:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - username
        - email
        - password

    ResumeScore:
      type: object
      properties:
        overall:
          type: number
        categories:
          type: object
          properties:
            formatting:
              type: number
            keywords:
              type: number
            experience:
              type: number
            education:
              type: number
            skills:
              type: number
          required:
            - formatting
            - keywords
            - experience
            - education
            - skills
        suggestions:
          type: array
          items:
            type: string
      required:
        - overall
        - categories
        - suggestions

    Job:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        company:
          type: string
        location:
          type: string
        salary:
          type: string
        matchScore:
          type: number
        description:
          type: string
        requirements:
          type: array
          items:
            type: string
        postedAt:
          type: string
        type:
          type: string
          enum: [Full-time, Part-time, Contract, Remote]
      required:
        - id
        - title
        - company
        - location
        - matchScore
        - description
        - type

    JobApplicationResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

security:
  - bearerAuth: []

paths:
  /auth/login:
    post:
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCredentials'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials

  /auth/signup:
    post:
      summary: User signup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupCredentials'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: User already exists

  /auth/logout:
    post:
      summary: User logout
      responses:
        '200':
          description: Logged out successfully

  /auth/me:
    get:
      summary: Get current user
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated

  /api/resume/analyze:
    post:
      summary: Analyze resume
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Analysis successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeScore'

  /api/jobs/recommendations:
    post:
      summary: Get job recommendations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeScore'
      responses:
        '200':
          description: List of recommended jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'

  /api/jobs/{id}/apply:
    post:
      summary: Apply to a job
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Job ID
      responses:
        '200':
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobApplicationResponse'